feat(view): プラン一覧を探すページに配置
