+++ src/pages/search.tsx
@@ -1,5 +1,25 @@
+import { Center, Spinner, VStack } from "@chakra-ui/react";
 import { GetStaticProps } from "next";
+import { useTranslation } from "next-i18next";
+import { serverSideTranslations } from "next-i18next/serverSideTranslations";
+import { useEffect } from "react";
+import { MdTrendingUp } from "react-icons/md";
+import InfiniteScroll from "react-infinite-scroller";
+import { PlannerGraphQlApi } from "src/data/graphql/PlannerGraphQlApi";
+import { createPlanFromPlanEntity } from "src/domain/factory/Plan";
+import { Plan } from "src/domain/models/Plan";
+import { RequestStatuses } from "src/domain/models/RequestStatus";
+import { PlannerApi } from "src/domain/plan/PlannerApi";
+import { hasValue } from "src/domain/util/null";
+import { TranslationNameSpaces, i18nAppConfig } from "src/locales/i18n";
+import {
+    fetchPlansRecentlyCreated,
+    pushPlansRecentlyCreated,
+    reduxPlanSelector,
+} from "src/redux/plan";
+import { useAppDispatch } from "src/redux/redux";
 import { Layout } from "src/view/common/Layout";
+import { Padding } from "src/view/constants/padding";
 import { Size } from "src/view/constants/size";
 import { useNearbyPlans } from "src/view/hooks/useNearbyPlans";
 import {
@@ -8,8 +28,22 @@ import {
 } from "src/view/navigation/BottomNavigation";
 import { NavBar } from "src/view/navigation/NavBar";
 import { NearbyPlanList } from "src/view/plan/NearbyPlanList";
+import { PlanList } from "src/view/plan/PlanList";
+import { PlanListSectionTitle } from "src/view/top/PlanListSectionTitle";
 
-export default function SearchPage() {
+type Props = {
+    plansRecentlyCreated: Plan[] | null;
+    nextPageTokenPlansRecentlyCreated: string | null;
+};
+
+export default function SearchPage(props: Props) {
+    const dispatch = useAppDispatch();
+    const { t } = useTranslation();
+    const {
+        plansRecentlyCreated,
+        nextPageTokenPlansRecentlyCreated,
+        fetchPlansRecentlyCreatedRequestStatus,
+    } = reduxPlanSelector();
     const {
         plansNearby,
         locationPermission,
@@ -17,6 +51,25 @@ export default function SearchPage() {
         isFetchingNearbyPlans,
         fetchNearbyPlans,
     } = useNearbyPlans();
+
+    useEffect(() => {
+        // すでにプランを取得済みの場合は何もしない
+        if (plansRecentlyCreated) return;
+
+        if (props.plansRecentlyCreated) {
+            // 初期表示時のみISRで取得したプランをReduxに保存する
+            dispatch(
+                pushPlansRecentlyCreated({
+                    plans: props.plansRecentlyCreated,
+                    nextPageTokenPlansRecentlyCreated:
+                        props.nextPageTokenPlansRecentlyCreated,
+                })
+            );
+        } else {
+            // ISRで取得できなかった場合はAPIから取得する
+            dispatch(fetchPlansRecentlyCreated());
+        }
+    }, [plansRecentlyCreated]);
 
     return (
         <Layout
@@ -26,21 +79,95 @@ export default function SearchPage() {
                 <BottomNavigation page={BottomNavigationPages.Search} />
             }
         >
-            {/* TODO: 拒否設定されている場合の対処をする */}
-            <NearbyPlanList
-                plans={plansNearby}
-                locationPermission={locationPermission}
-                px={Size.top.px}
-                isFetchingCurrentLocation={isFetchingCurrentLocation}
-                isFetchingNearbyPlans={isFetchingNearbyPlans}
-                onRequestFetchNearByPlans={fetchNearbyPlans}
-            />
+            <VStack w="100%" pb={Padding.p32}>
+                {/* TODO: 拒否設定されている場合の対処をする */}
+                <NearbyPlanList
+                    plans={plansNearby}
+                    locationPermission={locationPermission}
+                    px={Size.top.px}
+                    isFetchingCurrentLocation={isFetchingCurrentLocation}
+                    isFetchingNearbyPlans={isFetchingNearbyPlans}
+                    onRequestFetchNearByPlans={fetchNearbyPlans}
+                />
+                {/* TODO: React 18に対応 */}
+                {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
+                {/* @ts-ignore */}
+                <InfiniteScroll
+                    loadMore={() => dispatch(fetchPlansRecentlyCreated())}
+                    hasMore={hasValue(nextPageTokenPlansRecentlyCreated)}
+                    style={{ width: "100%" }}
+                >
+                    <PlanList plans={plansRecentlyCreated} px={Size.top.px}>
+                        <PlanListSectionTitle
+                            title={t("home:recentlyCreatedPlans")}
+                            icon={MdTrendingUp}
+                        />
+                    </PlanList>
+                    {fetchPlansRecentlyCreatedRequestStatus ===
+                        RequestStatuses.PENDING && (
+                        <Center w="100%" py="32px">
+                            <Spinner size="md" color="orange.600" />
+                        </Center>
+                    )}
+                </InfiniteScroll>
+            </VStack>
         </Layout>
     );
 }
 
-export const getStaticProps: GetStaticProps = async () => {
+export const getStaticProps: GetStaticProps<Props> = async ({ locale }) => {
+    const plannerApi: PlannerApi = new PlannerGraphQlApi();
+
+    let plans: Plan[] = [];
+    let nextPageKey: string | null = null;
+    try {
+        console.info({
+            page: "/",
+            message: "Fetching Recently Created Plans...",
+        });
+        const response = await plannerApi.fetchPlans({
+            pageKey: null,
+        });
+        plans = response.plans.map((entity) =>
+            createPlanFromPlanEntity(entity)
+        );
+        nextPageKey = response.nextPageKey;
+    } catch (e) {
+        console.error({
+            page: "/",
+            message: "Failed to fetch Recently Created Plans .",
+            error: e,
+        });
+
+        return {
+            props: {
+                plansRecentlyCreated: null,
+                nextPageTokenPlansRecentlyCreated: null,
+                ...(await serverSideTranslations(
+                    locale,
+                    TranslationNameSpaces,
+                    i18nAppConfig
+                )),
+            },
+            revalidate: 30,
+        };
+    } finally {
+        console.info({
+            page: "/",
+            message: "Fetched Recently Created Plans.",
+        });
+    }
+
     return {
-        props: {},
+        props: {
+            plansRecentlyCreated: plans,
+            nextPageTokenPlansRecentlyCreated: nextPageKey,
+            ...(await serverSideTranslations(
+                locale,
+                TranslationNameSpaces,
+                i18nAppConfig
+            )),
+        },
+        revalidate: 60,
     };
 };
